util restructure for PIMBench
