+++ PIMbench/Makefile
@@ -3,7 +3,7 @@
 # This file is licensed under the MIT License.
 # See the LICENSE file in the root of this repository for more details.
 
-SUBDIRS := $(wildcard */.)
+SUBDIRS := $(filter-out util/, $(wildcard */.))
 
 .PHONY: debug perf dramsim3_integ clean $(SUBDIRS)
 .DEFAULT_GOAL := perf
