+++ PIMbench/axpy/PIM/axpy.cpp
@@ -13,7 +13,7 @@
 #include <omp.h>
 #endif
 
-#include "../../util.h"
+#include "util.h"
 #include "libpimeval.h"
 
 using namespace std;
