+++ PIMbench/convolution/PIM/conv.cpp
@@ -11,7 +11,7 @@
 #if defined(_OPENMP)
 #include <omp.h>
 #endif
-#include "../../util.h"
+#include "util.h"
 #include <iomanip>
 #include <chrono>
