+++ PIMbench/knn/PIM/knn.cpp
@@ -20,7 +20,7 @@
 #include <omp.h>
 #endif
 
-#include "../../util.h"
+#include "util.h"
 #include "libpimeval.h"
 
 using namespace std;
