+++ PIMbench/triangle-count/PIM/tc.cpp
@@ -18,7 +18,7 @@
 #include <omp.h>
 #endif
 
-#include "../../util.h"
+#include "util.h"
 #include "libpimeval.h"
 
 #define DEBUG 0
