+++ PIMbench/vec-add/PIM/vec-add.cpp
@@ -12,7 +12,7 @@
 #include <omp.h>
 #endif
 
-#include "../../util.h"
+#include "util.h"
 #include "libpimeval.h"
 
 using namespace std;
